# swagger.yaml
swagger: '2.0'
info:
  title: Blog API
  description: API Documentation for my Blog
  termsOfService: https://github.com/okeneo/PersonalNest/blob/main/LICENSE
  contact:
    name: okenetega@gmail.com
  license:
    name: MIT License
  version: v1
host: localhost:8000
schemes:
- http
basePath: /blog
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Basic:
    type: basic
security:
- Basic: []
paths:
  /account/{username}/:
    get:
      operationId: account_read
      description: |-
        Get the data of a given user.
        Only users with the admin role are able to view the data available from this endpoint.
      parameters:
      - name: username
        in: path
        description: Username
        type: string
        required: true
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Account'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized Request
        '404':
          description: User Not Found
      tags:
      - account
    delete:
      operationId: account_delete
      description: Delete a user.
      parameters:
      - name: username
        in: path
        description: Username
        type: string
        required: true
      responses:
        '200':
          description: Successful Response
        '400':
          description: Bad Request
        '401':
          description: Unauthorized Request
        '404':
          description: User Not Found
      tags:
      - account
    parameters:
    - name: username
      in: path
      required: true
      type: string
...
definitions:
  Account:
    required:
    - username
    - email
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      last_login:
        title: Last login
        type: string
        format: date-time
        x-nullable: true
      is_superuser:
        title: Superuser status
        description: Designates that this user has all permissions without explicitly
          assigning them.
        type: boolean
      username:
        title: Username
        description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
          only.
        type: string
        pattern: ^[\w.@+-]+$
        maxLength: 150
        minLength: 1
      first_name:
        title: First name
        type: string
        maxLength: 150
      last_name:
        title: Last name
        type: string
        maxLength: 150
      is_staff:
        title: Staff status
        description: Designates whether the user can log into this admin site.
        type: boolean
      is_active:
        title: Active
        description: Designates whether this user should be treated as active. Unselect
          this instead of deleting accounts.
        type: boolean
      date_joined:
        title: Date joined
        type: string
        format: date-time
      role:
        title: Role
        type: string
        enum:
        - AUTHOR
        - ADMIN
        - READER
      bio:
        title: Bio
        type: string
        maxLength: 255
      email:
        title: Email address
        type: string
        format: email
        maxLength: 255
        minLength: 1
      is_email_verified:
        title: Is email verified
        type: boolean
      groups:
        description: The groups this user belongs to. A user will get all permissions
          granted to each of their groups.
        type: array
        items:
          description: The groups this user belongs to. A user will get all permissions
            granted to each of their groups.
          type: integer
        uniqueItems: true
      user_permissions:
        description: Specific permissions for this user.
        type: array
        items:
          description: Specific permissions for this user.
          type: integer
        uniqueItems: true
...